Move search results to use relative urls (#17411)
* Move search results to use relative urls

* ..and now we have real mark tags instead of em tags
